+++ src/motifAddQuantiles.py
@@ -166,15 +166,6 @@ def readTsv(fname):
 
             if tmpName not in patternIDs:
                 patternIDs.append(tmpName)
-
-    # Check for existing quantile information.
-    baseNames = ["contrib_magnitude", "seq_match", "contrib_match"]
-    quantileNames = [x + "_quantile" for x in baseNames]
-    for qn in quantileNames:
-        if qn in fieldNames:
-            # Weird, we already have a column for this data. Issue a warning.
-            logUtils.warning("The tsv {0:s} seems to already have quantile columns."
-                             .format(fname))
     return fieldNames, records, patternIDs
 
 
@@ -280,6 +271,15 @@ def main():
     logUtils.info("Reading in seqlet tsvs")
     seqletFields, seqletRecords, seqletPatternNames = readTsv(seqletInFname)
     scanFields, scanRecords, scanPatternNames = readTsv(scanInFname)
+    baseNames = ["contrib_magnitude", "seq_match", "contrib_match"]
+    quantileNames = [x + "_quantile" for x in baseNames]
+    # Check for existing quantile information.
+    for qn in quantileNames:
+        if qn in scanFields:
+            # Weird, we already have a column for this data. Issue a warning.
+            logUtils.warning("The tsv {0:s} seems to already have quantile columns."
+                             .format(scanInFname))
+            break
     logUtils.info("Finished reading.")
     readNames = ["contrib_magnitude", "seq_match", "contrib_match"]
     writeNames = [x + "_quantile" for x in readNames]